reset des valeurs dans lapage de création après la création, bouton "modifier" remplis les champs avec les valeurs de l'exercice sélectionné todo: faire en sorte qu'un exercice portant le même nom qu'un exercice existant écrase l'ancien(éviterais les doublons, moyen de modifier un exercice existant directement)

+++ MyOrthoOrtho/MyOrthoOrtho/ViewModels/ExerciceVM.cs
@@ -27,7 +27,7 @@ namespace MyOrthoOrtho.ViewModels
         public int PitchMin { get; set; }
         public int PitchMax { get; set; }
         public int IntensityThreshold { get; set; }
-        public int Duree_exacte { get; set; }
+        public decimal Duree_exacte { get; set; }
 
         public bool F0_exactEvaluated { get; set; }
         public bool F0_stableEvaluated { get; set; }
